[SW-71] refactor: 플랫폼별 설치 가이드를 개별 컴포넌트로 분리
