+++ src/components/guide/ios-not-safari.tsx
@@ -2,11 +2,11 @@
 
 import Logo from '@/components/ui/logo';
 
-export default function OtherGuide() {
+export default function NotSafariGuide() {
   return (
     <div className="flex flex-row items-center justify-center">
       <Logo size="base" />
-      <p>를 사용할 수 없는 장치입니다</p>
+      <p>를 사용하려면 Safari로 접속해주세요</p>
     </div>
   );
 }
